+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.7] - 2026-06-29
+
+### Fixed
+
+- **Legacy Recurring Event Rollover Announcements:** Fixed a bug where recurring events created before the bot was invited to a server would erroneously trigger a brand new "New Event" announcement precisely when they rolled over to their next occurrence.
+
 ## [1.6.6] - 2026-06-27
 
 ### Fixed